move ConsumerFactory class into `queues` module
