+++ streamtools/__init__.py
@@ -1,3 +1,3 @@
-from .queues import HandlerFactory, ConsumerRunner, CONSUMER_LOOPS, PRODUCERS
-from .factories import ConsumerFactory
+from .queues import HandlerFactory, ConsumerRunner, ConsumerFactory, \
+                    CONSUMER_LOOPS, PRODUCERS
 from .libs import prepare_aiohttp_app
